Moved ice_parsing to data_generators; updated to 1.1.7
