--- tensor2tensor/ice_parsing/__init__.py
@@ -1,2 +1,0 @@
-
-from .ice_parsing import IceParsingTokens, transformer_parsing_ice, transformer_parsing_ice_big
